Updated index form
Updated the index form to have requirements on the form.

Unfortunately the way that we have things laid out at the moment, HTML5 does not have a native way of grouping checkboxes and requiring that one is checked.
This means we need to either use Javascript or JQuery to validate the form.
I've been struggling to get any grouped form of the checkbox validation working, but, if I recall correctly, we are going to slim down the information per page anyway.

My proposed solution is that we think about moving over to a multiple form submission design instead of a singular form submission.  Although, I truthfully have no idea how we would do that.